# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443/
  description: Generated server url
paths:
...
  /api/entries/:
    get:
      tags:
      - entry-rest-controller
      operationId: getEntries
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry_Basic'
    post:
      tags:
      - entry-rest-controller
      operationId: createEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryDTO'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Entry'
...
  /api/comments/:
    get:
      tags:
      - comment-rest-controller
      operationId: getComments
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment_Comments'
    post:
      tags:
      - comment-rest-controller
      operationId: createComment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Comment'
  /api/auth/refresh:
    post:
      tags:
      - login-rest-controller
      operationId: refreshToken
      parameters:
      - name: refreshToken
        in: cookie
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /api/users:
    get:
      tags:
      - user-rest-controller
      summary: Get a list of the web app users.
      operationId: getUsers
      parameters:
      - name: firstname
        in: query
        description: Use this parameter to filter users by their name.
        required: false
        schema:
          type: string
      - name: surname
        in: query
        description: Use this parameter to filter users by their surname.
        required: false
        schema:
          type: string
      responses:
        "200":
          description: The list of users is displayed correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basic_Basic'
        "404":
          description: List of users not found
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      summary: Get a user by its id.
      operationId: getUser
      parameters:
      - name: id
        in: path
        description: This is the id of the user you're looking for.
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basic_Basic'
        "400":
          description: Invalid id supplied
        "404":
          description: User not found
    delete:
      tags:
      - user-rest-controller
      summary: Delete user by its id
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: Id of the user you want to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User deleted correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basic'
        "400":
          description: Invalid id supplied
        "404":
          description: User not found
...
  /api/entries/{id}/comments:
    get:
      tags:
      - entry-rest-controller
      operationId: getCommentsByIdEntry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Entry_EntryComments'
  /api/entries/{idEntry}:
    get:
      tags:
      - entry-rest-controller
      operationId: getEntry
      parameters:
      - name: idEntry
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Entry_Basic'
  /api/entries/{idEntry}/user:
    get:
      tags:
      - entry-rest-controller
      operationId: getUserByIdEntry
      parameters:
      - name: idEntry
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Entry_UserEntry'
...
  /api/comments/{id}:
    get:
      tags:
      - comment-rest-controller
      operationId: getComment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Comment_Comments'
...
components:
  schemas:
    Basic:
      type: object
...
    Comment:
      type: object
      properties:
        idComment:
          type: integer
          format: int64
        descriptionComment:
          type: string
        registrationDate:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        entry:
          $ref: '#/components/schemas/Entry'
    Criptocurrency:
      type: object
      properties:
        idCripto:
          type: integer
          format: int64
        nameCripto:
          type: string
        priceCripto:
          type: number
          format: double
        image:
          type: string
    Entry:
      type: object
      properties:
        idEntry:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        registrationDate:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        comment:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    EntryDTO:
      type: object
      properties:
        idEntry:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        registrationDate:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        idUser:
          type: integer
          format: int64
        encodedPassword:
          type: string
        name:
          type: string
        surname:
          type: string
        firstname:
          type: string
        email:
          type: string
        registrationDate:
          type: string
          format: date-time
        image:
          type: object
          properties:
            binaryStream:
              type: object
        follow:
          type: string
        tokenPass:
          type: string
        tokenCreationDate:
          type: string
          format: date-time
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        criptocurrencies:
          type: array
          items:
            $ref: '#/components/schemas/Criptocurrency'
        friends:
          type: array
          items:
            $ref: '#/components/schemas/User'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/Entry'
        roles:
          type: array
          items:
            type: string
    CommentDTO:
      type: object
      properties:
        idEntry:
          type: integer
          format: int64
        descriptionComment:
          type: string
    AuthResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILURE
        message:
          type: string
        error:
          type: string
...
    Basic_Basic:
      type: object
...
    Comment_EntryComments:
      type: object
      properties:
        idComment:
          type: integer
          format: int64
        descriptionComment:
          type: string
        registrationDate:
          type: string
          format: date-time
        entry:
          $ref: '#/components/schemas/Entry_EntryComments'
    Entry_EntryComments:
      type: object
      properties:
        idEntry:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        registrationDate:
          type: string
          format: date-time
        comment:
          type: array
          items:
            $ref: '#/components/schemas/Comment_EntryComments'
    Entry_Basic:
      type: object
      properties:
        idEntry:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        registrationDate:
          type: string
          format: date-time
    Entry_UserEntry:
      type: object
      properties:
        idEntry:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        registrationDate:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User_UserEntry'
    User_UserEntry:
      type: object
      properties:
        idUser:
          type: integer
          format: int64
        encodedPassword:
          type: string
        name:
          type: string
        surname:
          type: string
        firstname:
          type: string
        email:
          type: string
        registrationDate:
          type: string
          format: date-time
        image:
          type: object
          properties:
            binaryStream:
              type: object
        follow:
          type: string
        tokenPass:
          type: string
        tokenCreationDate:
          type: string
          format: date-time
        roles:
          type: array
          items:
            type: string
    Comment_Comments:
      type: object
      properties:
        idComment:
          type: integer
          format: int64
        descriptionComment:
          type: string
        registrationDate:
          type: string
          format: date-time
        entry:
          $ref: '#/components/schemas/Entry_Comments'
    Entry_Comments:
      type: object
...